+++ CourseWorkWF/Interface/ModelInterface/IProduct.cs
@@ -8,7 +8,7 @@ namespace CourseWorkWF.Interface.ModelInterface
 {
     public interface IProduct
     {
-        public int ProductID { get; set; }
+        public long ProductID { get; set; }
         public string ProductName { get; set; }
         public decimal Price { get; set; }
         public IProduct Clone();
