+++ CourseWorkWF/Interface/ModelInterface/ISell.cs
@@ -5,7 +5,7 @@ namespace CourseWorkWF.Interface.ModelInterface
 {
     public interface ISell 
     {
-        public IDictionary<int, IProductsCollectionItem> Products { get; set; }
+        public IDictionary<long, IProductsCollectionItem> Products { get; set; }
         public IMoneyOperation MoneyOperation { get; set; }
         public IDiscount Discount { get; set; }
     }
