+++ CourseWorkWF/Interface/ViewInterface/IRefundFormView.cs
@@ -15,7 +15,7 @@ namespace CourseWorkWF.Interface.ViewInterface
         public int SelectedReceiptNumber { get; }
         public DateTime DateFrom { get; }
         public DateTime DateTo { get; }
-        public int ProductID { get; }       
+        public long ProductID { get; }       
         public OperationMethod OperationType { get; }
         public int Amount { get; }
         public string Reason { get; }
