+++ CourseWorkWF/Models/Refund.cs
@@ -7,14 +7,14 @@ namespace CourseWorkWF.Models
     public class Refund : IRefund,IRevenueChanger
     {
         private string _reason;
-        private IDictionary<int, IProductsCollectionItem> _products;
+        private IDictionary<long, IProductsCollectionItem> _products;
         private IMoneyOperation _moneyOperation;
         public IMoneyOperation MoneyOperation 
         {
             get { return _moneyOperation; }
             set { if (value == null) throw new ArgumentNullException("MoneyOperation не может быть null"); _moneyOperation = value; }
         }
-        public IDictionary<int, IProductsCollectionItem> Products
+        public IDictionary<long, IProductsCollectionItem> Products
         {
             get { return _products; }
             set { if (value == null) throw new ArgumentNullException("Products не может быть null"); _products = value; }
@@ -25,7 +25,7 @@ namespace CourseWorkWF.Models
             set { if (string.IsNullOrEmpty(value)) throw new ArgumentException("Reason не может быть пустой"); _reason = value; }
         }
 
-        public Refund(IDictionary<int, IProductsCollectionItem> products, IMoneyOperation moneyOperation, string reason)
+        public Refund(IDictionary<long, IProductsCollectionItem> products, IMoneyOperation moneyOperation, string reason)
         {
             Products = products;
             MoneyOperation = moneyOperation;
