+++ CourseWorkWF/Models/Sell.cs
@@ -6,10 +6,10 @@ namespace CourseWorkWF.Models
 {
     public class Sell : ISell, IRevenueChanger
     {
-        private IDictionary<int, IProductsCollectionItem> _products;
+        private IDictionary<long, IProductsCollectionItem> _products;
         private IDiscount _discount;
         public IMoneyOperation MoneyOperation { get; set; }
-        public IDictionary<int, IProductsCollectionItem> Products 
+        public IDictionary<long, IProductsCollectionItem> Products 
         { 
             get { return _products; }
             set { if (value == null) throw new ArgumentNullException("Products не может быть null"); _products = value; }
@@ -21,7 +21,7 @@ namespace CourseWorkWF.Models
             set { if (value == null) throw new ArgumentNullException("Discount не может быть null"); _discount = value; }
         }
 
-        public Sell(IDictionary<int, IProductsCollectionItem> products, IMoneyOperation moneyOperation, IDiscount discount)
+        public Sell(IDictionary<long, IProductsCollectionItem> products, IMoneyOperation moneyOperation, IDiscount discount)
         {
             Products = products;
             MoneyOperation = moneyOperation;
